+++ src/shared/protocol.rs
@@ -78,6 +78,16 @@ impl ProtocolHasher {
         self.hash::<E>(ProtocolPart::ClientEvent);
     }
 
+    pub(crate) fn add_broadcast_message<E>(&mut self) {
+        debug!("adding broadcast message `{}`", ShortName::of::<E>());
+        self.hash::<E>(ProtocolPart::BroadcastMessage);
+    }
+
+    pub(crate) fn add_broadcast_event<E>(&mut self) {
+        debug!("adding broadcast event `{}`", ShortName::of::<E>());
+        self.hash::<E>(ProtocolPart::BroadcastEvent);
+    }
+
     pub(crate) fn add_server_message<E>(&mut self) {
         debug!("adding server message `{}`", ShortName::of::<E>());
         self.hash::<E>(ProtocolPart::ServerMessage);
@@ -133,6 +143,8 @@ enum ProtocolPart {
     IndependentMessage,
     IndependentEvent,
     TrackMutateMessages,
+    BroadcastMessage,
+    BroadcastEvent,
 }
 
 /// Hash of all registered events and replication rules.