+++ tests/broadcast_message.rs
@@ -1,0 +1,195 @@
+use bevy::{ecs::entity::MapEntities, prelude::*, state::app::StatesPlugin, time::TimePlugin};
+use bevy_replicon::{
+    prelude::*,
+    shared::server_entity_map::ServerEntityMap,
+    test_app::{ServerTestAppExt, TestClientEntity},
+};
+use serde::{Deserialize, Serialize};
+use test_log::test;
+
+#[test]
+fn regular() {
+    let mut server_app = App::new();
+    let mut client_app = App::new();
+    for app in [&mut server_app, &mut client_app] {
+        app.add_plugins((MinimalPlugins, StatesPlugin, RepliconPlugins))
+            .add_broadcast_message::<Test>(Channel::Ordered)
+            .finish();
+    }
+
+    server_app.connect_client(&mut client_app);
+    let client_entity = **client_app.world().resource::<TestClientEntity>();
+
+    client_app.world_mut().write_message(Test);
+
+    client_app.update();
+
+    let local = client_app.world().resource::<Messages<Broadcast<Test>>>();
+    let mut cursor = local.get_cursor();
+    let broadcasts: Vec<_> = cursor.read(local).collect();
+    assert_eq!(broadcasts.len(), 1);
+    assert!(matches!(broadcasts[0].broadcaster, Broadcaster::Local));
+
+    server_app.exchange_with_client(&mut client_app);
+    server_app.update();
+
+    let received = server_app.world().resource::<Messages<Broadcast<Test>>>();
+    let mut cursor = received.get_cursor();
+    let broadcasts: Vec<_> = cursor.read(received).collect();
+    assert_eq!(broadcasts.len(), 1);
+    assert!(matches!(
+        broadcasts[0].broadcaster,
+        Broadcaster::Remote(ClientId::Client(entity)) if entity == client_entity
+    ));
+}
+
+#[test]
+fn mapped() {
+    let mut server_app = App::new();
+    let mut client_app = App::new();
+    for app in [&mut server_app, &mut client_app] {
+        app.add_plugins((
+            MinimalPlugins,
+            StatesPlugin,
+            RepliconPlugins.set(ServerPlugin::new(PostUpdate)),
+        ))
+        .add_mapped_broadcast_message::<WithEntity>(Channel::Ordered)
+        .finish();
+    }
+
+    server_app.connect_client(&mut client_app);
+
+    let server_entity = server_app.world_mut().spawn(Replicated).id();
+
+    server_app.update();
+    server_app.exchange_with_client(&mut client_app);
+    client_app.update();
+
+    let client_entity = *client_app
+        .world()
+        .resource::<ServerEntityMap>()
+        .to_client()
+        .get(&server_entity)
+        .unwrap();
+
+    client_app
+        .world_mut()
+        .write_message(WithEntity(client_entity));
+
+    client_app.update();
+
+    let local = client_app
+        .world()
+        .resource::<Messages<Broadcast<WithEntity>>>();
+    let mut cursor = local.get_cursor();
+    let local_entities: Vec<_> = cursor.read(local).map(|m| m.0).collect();
+    assert_eq!(local_entities, [client_entity]);
+
+    server_app.exchange_with_client(&mut client_app);
+    server_app.update();
+
+    let received = server_app
+        .world()
+        .resource::<Messages<Broadcast<WithEntity>>>();
+    let mut cursor = received.get_cursor();
+    let received_entities: Vec<_> = cursor.read(received).map(|m| m.0).collect();
+    assert_eq!(received_entities, [server_entity]);
+}
+
+#[test]
+fn without_plugins() {
+    let mut server_app = App::new();
+    let mut client_app = App::new();
+    server_app
+        .add_plugins((
+            MinimalPlugins,
+            StatesPlugin,
+            RepliconPlugins
+                .build()
+                .disable::<ClientPlugin>()
+                .disable::<ClientMessagePlugin>(),
+        ))
+        .add_broadcast_message::<Test>(Channel::Ordered)
+        .finish();
+    client_app
+        .add_plugins((
+            MinimalPlugins,
+            StatesPlugin,
+            RepliconPlugins
+                .build()
+                .disable::<ServerPlugin>()
+                .disable::<ServerMessagePlugin>(),
+        ))
+        .add_broadcast_message::<Test>(Channel::Ordered)
+        .finish();
+
+    server_app.connect_client(&mut client_app);
+
+    client_app.world_mut().write_message(Test);
+
+    client_app.update();
+    server_app.exchange_with_client(&mut client_app);
+    server_app.update();
+
+    let received = server_app.world().resource::<Messages<Broadcast<Test>>>();
+    let mut cursor = received.get_cursor();
+    let broadcasts: Vec<_> = cursor.read(received).collect();
+    assert_eq!(broadcasts.len(), 1);
+    assert!(matches!(
+        broadcasts[0].broadcaster,
+        Broadcaster::Remote(ClientId::Client(_))
+    ));
+}
+
+#[test]
+fn local_sending() {
+    let mut app = App::new();
+    app.add_plugins((TimePlugin, StatesPlugin, RepliconPlugins))
+        .add_broadcast_message::<Test>(Channel::Ordered)
+        .finish();
+
+    app.world_mut().write_message(Test);
+
+    app.update();
+
+    let messages = app.world().resource::<Messages<Test>>();
+    assert!(messages.is_empty());
+
+    let broadcasts = app.world().resource::<Messages<Broadcast<Test>>>();
+    let mut cursor = broadcasts.get_cursor();
+    let collected: Vec<_> = cursor.read(broadcasts).collect();
+    assert_eq!(collected.len(), 1);
+    assert!(matches!(collected[0].broadcaster, Broadcaster::Local));
+}
+
+#[test]
+fn with_disconnect() {
+    let mut server_app = App::new();
+    let mut client_app = App::new();
+    for app in [&mut server_app, &mut client_app] {
+        app.add_plugins((MinimalPlugins, StatesPlugin, RepliconPlugins))
+            .add_broadcast_message::<Test>(Channel::Ordered)
+            .finish();
+    }
+
+    server_app.connect_client(&mut client_app);
+
+    client_app.world_mut().write_message(Test);
+
+    server_app.disconnect_client(&mut client_app);
+
+    let messages = client_app.world().resource::<Messages<Test>>();
+    assert!(messages.is_empty());
+
+    let broadcasts = client_app.world().resource::<Messages<Broadcast<Test>>>();
+    assert!(
+        broadcasts.is_empty(),
+        "client shouldn't resend broadcasts locally after disconnect"
+    );
+}
+
+#[derive(Deserialize, Message, Serialize)]
+struct Test;
+
+#[derive(Deserialize, Message, Serialize, Clone, MapEntities)]
+struct WithEntity(#[entities] Entity);